+++ integration-test/lincheck/src/main/org/jetbrains/lincheck_test/datastructures/linearizable/ConcurrentHashMapTest.kt
@@ -32,6 +32,7 @@ class ConcurrentHashMapTest : AbstractLincheckTest() {
     fun remove(@Param(name = "key") key: Int) = map.remove(key)
 
     override fun <O : Options<O, *>> O.customize() {
+        // TODO: This times out, we need to fix performance
         ignoreExperimentalModelChecking()
         // To obtain rare interleaving with `fullAddCount` method
         if (this is ModelCheckingOptions) {
