+++ lincheck/src/jvm/main/org/jetbrains/lincheck/Lincheck.kt
@@ -46,9 +46,8 @@ object Lincheck {
     @JvmOverloads
     fun runConcurrentTest(
         invocations: Int = DEFAULT_INVOCATIONS,
-        useExperimentalModelChecking: Boolean = false,
         block: Runnable
-    ) = runConcurrentTestInternal(invocations, useExperimentalModelChecking, LincheckSettings.DEFAULT, block)
+    ) = runConcurrentTestInternal(invocations, LincheckSettings.DEFAULT, block)
 
     /**
      * This method will explore different interleavings of the [block] body and all the threads created within it,
@@ -63,7 +62,6 @@ object Lincheck {
     @JvmStatic
     internal fun runConcurrentTestInternal(
         invocations: Int = DEFAULT_INVOCATIONS,
-        useExperimentalModelChecking: Boolean = false,
         settings: LincheckSettings,
         block: Runnable
     ) {
@@ -73,7 +71,7 @@ object Lincheck {
             .recursionBound(settings.recursionBound)
             .loopIterationsBeforeThreadSwitch(settings.loopIterationsBeforeThreadSwitch)
 
-        if(useExperimentalModelChecking)
+        if(settings.useExperimentalModelChecking)
             options = options.useExperimentalModelChecking()
 
         val testCfg = options.createTestConfigurations(block::class.java)
