+++ lincheck/src/jvm/test/org/jetbrains/kotlinx/lincheck_test/strategy/eventstructure/PrimitivesTest.kt
@@ -37,7 +37,6 @@ import org.junit.rules.TestName
 import kotlin.reflect.jvm.javaMethod
 import org.jetbrains.lincheck.util.UnsafeHolder
 import kotlin.concurrent.thread
-import java.lang.reflect.Array as ArrayReflection
 
 class PrimitivesTest {
 
@@ -1518,335 +1517,6 @@ class PrimitivesTest {
         }
     }
 
-
-    @Test
-    fun testGC() {
-        class Box(val x: Int) {}
-        litmusTest(assertSame(setOf(null, 0, 42_000_000), UNKNOWN)) {
-            var x: Box? = Box(0)
-            var r0 : Int? = -1
-
-            val t1 = thread {
-                System.gc()
-                System.gc()
-                x = Box(42_000_000)
-                System.gc()
-                System.gc()
-            }
-            val t2 = thread {
-                System.gc()
-                System.gc()
-                x = null
-                System.gc()
-                System.gc()
-            }
-            val t3 = thread {
-                System.gc()
-                System.gc()
-                r0 = x?.x
-                System.gc()
-                System.gc()
-            }
-
-            t1.join()
-            t2.join()
-            t3.join()
-            r0
-        }
-    }
-
-    @Test
-    fun testArrayCopy() {
-        val outcomes = setOf(
-            listOf(1,2,3,4),
-            listOf(1,2,3,0),
-            listOf(1,2,0,0),
-            listOf(1,0,0,0),
-            listOf(0,0,0,0),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-            val a = IntArray(N)
-            val b = IntArray(N)
-            val result = IntArray(N)
-
-            for (i in 0 until N) {
-                a[i] = i + 1
-            }
-
-            val t1 = thread {
-                System.arraycopy(a, 0, b, 0, N)
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    result[i] = b[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            result.toList()
-        }
-    }
-
-    @Test
-    fun testArrayCopyStrings() {
-        val outcomes = setOf(
-            listOf("1","2","3","4"),
-            listOf("1","2","3","0"),
-            listOf("1","2","0","0"),
-            listOf("1","0","0","0"),
-            listOf("0","0","0","0"),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-            val a = Array<String>(N, { "0" })
-            val b = Array<String>(N, { "0" })
-            val result = Array<String>(N, { "0" })
-
-            for (i in 0 until N) {
-                a[i] = "${i + 1}"
-            }
-
-            val t1 = thread {
-                System.arraycopy(a, 0, b, 0, N)
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    result[i] = b[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            result.toList()
-        }
-    }
-
-
-    @Test
-    fun testArrayCopyBoxes() {
-        class Box(val x: Int) {
-            override fun toString(): String {
-                return "$x"
-            }
-            override fun hashCode(): Int {
-                return x
-            }
-            override fun equals(other: Any?): Boolean {
-                return other is Box && other.x == x
-            }
-        }
-
-        val outcomes = setOf(
-            listOf(Box(1),Box(2),Box(3),Box(4)),
-            listOf(Box(1),Box(2),Box(3),Box(0)),
-            listOf(Box(1),Box(2),Box(0),Box(0)),
-            listOf(Box(1),Box(0),Box(0),Box(0)),
-            listOf(Box(0),Box(0),Box(0),Box(0)),
-        )
-
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-            val a = Array<Box>(N, { Box(0) })
-            val b = Array<Box>(N, { Box(0) })
-            val result = Array<Box>(N, { Box(0) })
-
-            for (i in 0 until N) {
-                a[i] = Box(i + 1)
-            }
-
-            val t1 = thread {
-                System.arraycopy(a, 0, b, 0, N)
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    result[i] = b[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            result.toList()
-        }
-    }
-
-
-    @Test
-    fun testArrayNewInstancePrimitives() {
-        val outcomes = setOf(
-            listOf(1,2,3,4),
-            listOf(1,2,3,0),
-            listOf(1,2,0,0),
-            listOf(1,0,0,0),
-            listOf(0,0,0,0),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-            val arr = ArrayReflection.newInstance(Int::class.javaPrimitiveType, N) as IntArray
-            val res = IntArray(N)
-
-
-            val t1 = thread {
-                for(i in 0 until N) {
-                    arr[i] = i + 1
-                }
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    res[i] = arr[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            res.toList()
-        }
-    }
-
-    @Test
-    fun testArrayNewInstanceStrings() {
-        val outcomes = setOf(
-            listOf("1" , "2" , "3" , "4"),
-            listOf("1" , "2" , "3" , null),
-            listOf("1" , "2" , null, null),
-            listOf("1" , null, null, null),
-            listOf(null, null, null, null),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-
-            @Suppress("UNCHECKED_CAST")
-            val arr = ArrayReflection.newInstance(String::class.java, N) as Array<String?>
-            val res = Array<String?>(N, { "0" })
-
-
-            val t1 = thread {
-                for(i in 0 until N) {
-                    arr[i] = "${i + 1}"
-                }
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    res[i] = arr[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            res.toList()
-        }
-    }
-
-    @Test
-    fun testArrayNewInstanceBoxes() {
-
-        class Box(val x: Int) {
-            override fun toString(): String {
-                return "$x"
-            }
-            override fun hashCode(): Int {
-                return x
-            }
-            override fun equals(other: Any?): Boolean {
-                return other is Box && other.x == x
-            }
-        }
-
-        val outcomes = setOf(
-            listOf(Box(1) , Box(2) , Box(3) , Box(4)),
-            listOf(Box(1) , Box(2) , Box(3) , null),
-            listOf(Box(1) , Box(2) , null, null),
-            listOf(Box(1) , null, null, null),
-            listOf(null, null, null, null),
-        )
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-
-            @Suppress("UNCHECKED_CAST")
-            val arr = ArrayReflection.newInstance(Box::class.java, N) as Array<Box?>
-            val res = Array<Box?>(N, { Box(0) })
-
-
-            val t1 = thread {
-                for(i in 0 until N) {
-                    arr[i] = Box(i+1)
-                }
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    res[i] = arr[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            res.toList()
-        }
-    }
-
-    @Test
-    fun testArrayNewInstanceArrays() {
-
-        class Box(val x: Int) {
-            override fun toString(): String {
-                return "$x"
-            }
-            override fun hashCode(): Int {
-                return x
-            }
-            override fun equals(other: Any?): Boolean {
-                return other is Box && other.x == x
-            }
-        }
-
-        val outcomes = setOf<List<Box?>>(
-            listOf(Box(1) , Box(2) , Box(3) , Box(4)),
-            listOf(Box(1) , Box(2) , Box(3) , null),
-            listOf(Box(1) , Box(2) , null , null),
-            listOf(Box(1) , null , null, null ),
-            listOf(null, null, null, null),
-        )
-
-        litmusTest(assertSame(outcomes)) {
-            val N = 4
-
-            @Suppress("UNCHECKED_CAST")
-            val arr = ArrayReflection.newInstance(Array<Box?>::class.java, N) as Array<Array<Box?>?>
-            val res = Array<Array<Box?>?>(N, { arrayOf(Box(0)) })
-
-            val t1 = thread {
-                for(i in 0 until N) {
-                    arr[i] = arrayOf(Box(i+1))
-                }
-            }
-
-            val t2 = thread {
-                for (i in N-1 downTo  0) {
-                    res[i] = arr[i]
-                }
-            }
-
-            t1.join()
-            t2.join()
-
-            res.map { if (it == null) null else it[0] }.toList()
-        }
-    }
-
-
     @Test
     fun testFunctionInConstructor() {
         class Box(var value: Int) {}
@@ -1971,7 +1641,7 @@ class PrimitivesTest {
                 actor(TestClass::post)
             }
         }
-        val outcomes =  setOf(5,7)
+        val outcomes = setOf(5,7)
 
         litmusTest(TestClass::class.java, testScenartio, assertSame(outcomes, UNKNOWN)) { results ->
             val b1 = getValue<Int>(results.postResults[0]!!)
